refactor: move missions transformation to release repository
